blog application api

+++ src/main/java/com/vm/Vyas_Media/Controller/MovetoTrashRestController.java
@@ -1,7 +1,6 @@
 package com.vm.Vyas_Media.Controller;
 
 import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -11,10 +10,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-
 import com.vm.Vyas_Media.Entity.MovetoTrash;
 import com.vm.Vyas_Media.Service.MovetoTrashService;
-
 
 @RestController
 @RequestMapping("/users/rest")
@@ -29,7 +26,7 @@ public class MovetoTrashRestController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public void delete(@PathVariable Long id) {
+    public void delete(@PathVariable Integer id) {
         userService.delete(id);
     }
 }
